refactor and added app version name
